Fix interger overflow in Canon exposure calculation.

+++ ufraw_developer.c
@@ -342,7 +342,7 @@ void developer_prepare(developer_data *d, conf_data *conf,
     unsigned exposure = pow(2, conf->exposure) * 0x10000;
     /* Handle the exposure normalization for Canon EOS cameras. */
     if ( conf->ExposureNorm>0 )
-	exposure = exposure * d->rgbMax / conf->ExposureNorm;
+	exposure = (guint64)exposure * d->rgbMax / conf->ExposureNorm;
     if ( exposure>=0x10000 ) d->restoreDetails = clip_details;
     if ( exposure<=0x10000 ) clipHighlights = digital_highlights;
     /* Check if gamma curve data has changed. */
